requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.2
attrs==24.2.0
auth0-python==4.7.2
autobahn==23.6.1
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
blinker==1.6.2
blosc==1.11.2
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.3.0
certifi==2024.8.30
cffi==1.15.1
cfgv==3.4.0
chardet==5.1.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
cohere==5.11.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
coverage==7.2.3
crewai==0.79.4
crewai-tools==0.14.0
cryptography==43.0.3
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.3.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo-search==2.8.5
durationpy==0.9
embedchain==0.1.125
executing==1.2.0
fastapi==0.115.4
fastavro==1.9.7
fastjsonschema==2.17.1
filelock==3.16.1
flake8==6.0.0
FLAML==2.3.2
Flask==2.3.2
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.10.0
gast==0.6.0
google-api-core==2.23.0
google-api-python-client==2.85.0
google-auth==2.36.0
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
grpcio-tools==1.67.1
gTTS==2.3.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
identify==2.6.1
idna==3.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.6.3
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.1
json5==0.9.14
json_repair==0.30.1
jsonpatch==1.33
jsonpickle==3.4.2
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_proxy==4.0.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
kaggle==1.6.17
keras==3.5.0
kiwisolver==1.4.4
kubernetes==31.0.0
lancedb==0.15.0
langchain==0.3.7
langchain-cohere==0.3.1
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langsmith==0.1.142
libclang==18.1.1
lightgbm==4.5.0
litellm==1.52.4
loguru==0.7.0
lxml==4.9.2
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.23.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.1.29
miceforest==5.7.0
mistune==2.0.5
ml-dtypes==0.3.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.8
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.4.2
nodeenv==1.9.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.3.1
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pexpect==4.8.0
pfzy==0.3.4
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==4.3.6
playsound==1.2.2
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
pre_commit==4.0.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
py==1.11.0
pyarrow==18.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyautogen==0.4
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydicom==3.0.1
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.9.0
pylance==0.19.1
pyparsing==3.0.9
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.387
pyrsistent==0.19.3
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv
python-json-logger==2.0.7
python-rapidjson==1.20
python-slugify==8.0.4
pytorch-minimize==0.0.2
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==25.1.0
qdrant-client==1.12.1
qtconsole==5.4.3
QtPy==2.3.1
ratelimiter==1.2.0.post0
readability-lxml==0.8.1
redis==4.5.4
referencing==0.35.1
regex==2024.11.6
requests>=2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.21.0
rsa==4.9
scapy==2.4.5
schema==0.7.7
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.26.1
semver==3.0.2
Send2Trash==1.8.2
shapely==2.0.6
shellingham==1.5.4
simpervisor==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.35
stack-data==0.6.2
starlette==0.41.2
swarm @ git+ssh://git@github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.1.0
tomli_w==1.1.0
torch==2.2.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.27.0
trio-websocket==0.11.1
tritonclient==2.51.0
txaio==23.1.1
typer==0.9.4
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.2.3
uv==0.5.1
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.26.5
vpython==7.6.4
watchfiles==0.24.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
zipp==3.20.2
boto3
paramiko>=3.4.0
